backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
langgraph>=0.2
google-generativeai>=0.8
python-dotenv>=1.0
pydantic>=2.0
